feature: partial redesign and added a lists page

+++ src/components/card/index.tsx
@@ -1,3 +1,4 @@
+import type { ReactNode } from 'react';
 import classnames from 'classnames';
 import { FaTrash, FaClone } from 'react-icons/fa';
 import { Card as MatterCard, Stack, IconButton } from '@fjlaubscher/matter';
@@ -6,15 +7,27 @@ import type { CardProps } from '@fjlaubscher/matter';
 import styles from './card.module.scss';
 
 type Props = Omit<CardProps, 'children'> & {
-  children?: React.ReactNode;
+  children?: ReactNode;
+  description?: string;
   onCopyClick?: () => void;
   onDeleteClick?: () => void;
 };
 
-const Card = ({ children, className, onClick, onCopyClick, onDeleteClick, title }: Props) => (
+const Card = ({
+  children,
+  className,
+  onClick,
+  onCopyClick,
+  onDeleteClick,
+  title,
+  description
+}: Props) => (
   <MatterCard className={classnames(styles.card, className)} onClick={onClick}>
-    <Stack className={styles.title} direction="row">
-      <span>{title}</span>
+    <Stack className={styles.header} direction="row">
+      <Stack className={styles.title} direction="column">
+        <span>{title}</span>
+        {description && <span className={styles.description}>{description}</span>}
+      </Stack>
       {onCopyClick || onDeleteClick ? (
         <Stack className={styles.buttons} direction="row">
           {onCopyClick && (
@@ -23,7 +36,9 @@ const Card = ({ children, className, onClick, onCopyClick, onDeleteClick, title 
               variant="error"
               onClick={(e) => {
                 e.stopPropagation();
-                onCopyClick();
+                if (onCopyClick) {
+                  onCopyClick();
+                }
               }}
             >
               <FaClone />
@@ -35,7 +50,9 @@ const Card = ({ children, className, onClick, onCopyClick, onDeleteClick, title 
               variant="error"
               onClick={(e) => {
                 e.stopPropagation();
-                onDeleteClick();
+                if (onDeleteClick) {
+                  onDeleteClick();
+                }
               }}
             >
               <FaTrash />
